requirements.txt
numpy
pandas
matplotlib
scikit-learn
Pillow
torch
torchvision
diffusers